LVS working on SKP branch
 Changes to be committed:
	modified:   efpmd/src/torch.c
	modified:   efpmd/torch/ANIModel.h
	modified:   efpmd/torch/c_libtorch.cc
	modified:   efpmd/torch/c_libtorch.h
	new file:   nnlib/ANIMM.pt
	new file:   nnlib/ANIMM_custom.pt
	modified:   setup.sh
	new file:   tests/aev_scripted.pt
	new file:   tests/ec.dat
	new file:   tests/grad.dat
	new file:   tests/gtest_elpot.in
	new file:   tests/lig.efp
	new file:   tests/sol.efp
	new file:   tests/test_libefp.inp

+++ efpmd/torch/c_libtorch.h
@@ -51,7 +51,10 @@ void load_custom_ani_model(ANIModel* model, const char* aev_name, const char* mo
 void get_ani_energy_grad(ANIModel* model, float* coordinates, int* species, double* ani_energy, float* gradients, float* forces, int num_atoms, int print);
 //void get_ani_energy_grad(ANIModel* model, float* coordinates, int* species, float* atomic_energies, float* gradients, float* forces, int num_atoms); 
 //void get_custom_energy_grad_wrapper(ANIModel* model, float* coordinates, int64_t* species, float* elecpots, int num_atoms, float* custom_energy, float* gradients, float* forces);
-void get_custom_energy_grad_wrapper(ANIModel* model, float* coordinates, int64_t* species, float* elecpots, int num_atoms, double* custom_energy, float* gradients, float* forces, int print);
+//void get_custom_energy_grad_wrapper(ANIModel* model, float* coordinates, int64_t* species, float* elecpots, int num_atoms, double* custom_energy, float* gradients, float* forces, int print);
+
+void get_custom_energy_grad_wrapper(ANIModel* model, double* coordinates, int64_t* species, double* elecpots, int num_atoms, double* custom_energy, double* gradients, double* forces, int print);
+
  
 void ANIModel_delete(ANIModel* model);
